+++ Common/CNLIB/Socket.h
@@ -35,7 +35,7 @@ public:
 
 	struct Hash
 	{
-		size_t operator()(const Socket& sock) const
+		UINT operator()(const Socket& sock) const
 		{
 			return std::hash<SOCKET>()(sock.pc);
 		}
